useJump and definitions
